+++ Get-FolderAge.ps1
@@ -1,6 +1,6 @@
 <#PSScriptInfo
 
-.VERSION 1.1
+.VERSION 1.2
 .GUID c9788cc2-d4af-4219-bf7d-8dd8fa89584f
 .AUTHOR Igor Iric, iricigor@gmail.com, https://github.com/iricigor
 .COMPANYNAME 
@@ -12,7 +12,7 @@
 .EXTERNALMODULEDEPENDENCIES 
 .REQUIREDSCRIPTS 
 .EXTERNALSCRIPTDEPENDENCIES 
-.RELEASENOTES Added -Exclude and -Threads parameters, for more info see https://github.com/iricigor/GetFolderAge/blob/master/ReleaseNotes.md
+.RELEASENOTES Added restartable script functionality and alias, bug fixes for append and threads issues, for more info see https://github.com/iricigor/GetFolderAge/blob/master/ReleaseNotes.md
 .DESCRIPTION Get-FolderAge returns `LastModifiedDate` for a specified folder(s) and if folders were modified after a specified cut-off date.
 
 #>
@@ -103,8 +103,11 @@ function Global:Get-FolderAge {
     If both CutOffTime and CutOffDays specified, the script will throw a warning.
     
     .PARAMETER OutputFile
-    A string specifying file name which will be used for output. If not specified, there will be no file output generated.
+    A string specifying file name which will be used for output in addition to screen (or pipeline) output.
     This is especially useful for long running commands. Each folder as soon as processed will be stored in the file.
+    This can be also used for restarting the script, if it gets interrupted before it finishes all folders.
+    Just specify the same input and output files, and script will skip already processed folders!
+    If this parameter is not specified, there will be no file output generated.
     
     .PARAMETER Exclude
     Specifies, as a string array, an folder names that this cmdlet excludes in the search operation.
@@ -163,7 +166,8 @@ function Global:Get-FolderAge {
 
         [parameter(Mandatory=$false)] [string[]]$Exclude,
 
-        [parameter(Mandatory=$false)] [int]$Threads = 0,
+        [parameter(Mandatory=$false)]
+        [ValidateRange(0,50)]         [int]$Threads = 0,
 
         #
         # Switches
@@ -185,10 +189,10 @@ function Global:Get-FolderAge {
 
         # process $InputFile
         if ($InputFile) {
-            if (!(Test-Path $InputFile)) {
+            if (!(Test-Path -LiteralPath $InputFile)) {
                 throw "$FunctionName cannot find input file $InputFile"
             }
-            $FolderName = Get-Content -Path $InputFile -ErrorAction SilentlyContinue
+            $FolderName = Get-Content -LiteralPath $InputFile -ErrorAction SilentlyContinue
             if ($FolderName) {
                 Write-Verbose -Message "$(Get-Date -f T)   successfully read $InputFile with $(@($FolderName).Count) entries"
             } else {
@@ -209,7 +213,6 @@ function Global:Get-FolderAge {
             }
         }
 
-        $First = $true # Used if there is output to file, only first line drops header
         $Separator = [IO.Path]::DirectorySeparatorChar
         $UC = '\\?\'
 
@@ -240,6 +243,34 @@ function Global:Get-FolderAge {
             }
 
         }
+
+        # Restartable script setup
+        if ($OutputFile) {
+            if (Test-Path -LiteralPath $OutputFile) {
+                $RP = Resolve-Path -LiteralPath $OutputFile
+                try {
+                    $FoldersToSkip = Import-Csv -LiteralPath $OutputFile | Select -Expand Path
+                    Write-Verbose -Message "$(Get-Date -f T)   Script continues writing to $OutputFile, with skipping $(@($FoldersToSkip).Count) processed folders"
+                } catch {
+                    throw "$FunctionName found existing file $OutputFile in unrecognized format, cannot continue."
+                }
+            } else {
+                $FoldersToSkip = $null
+                try {
+                    New-Item $OutputFile -ItemType File -ea Stop | Out-Null
+                    $RP = Resolve-Path -LiteralPath $OutputFile
+                    Remove-Item $OutputFile -Force
+                } catch {
+                    throw "$FunctionName cannot write output file $OutputFile`: $_"
+                }
+            }            
+            if ($RP.Provider.Name -ne 'FileSystem') {
+                throw "$FunctionName provided output file $OutputFile is not on the FileSystem"
+            } elseif ($OutputFile -ne $RP.ProviderPath) {
+                Write-Verbose -Message "$(Get-Date -f T)   Expanding $OutputFile to $($RP.Path) via $($RP.Provider.Name) call"
+                $OutputFile = $RP.ProviderPath
+            }
+        }
     }
 
 
@@ -253,7 +284,7 @@ function Global:Get-FolderAge {
                 Write-Error "$FunctionName cannot find folder $FolderEntry"
                 continue
             }
-            $RP = Resolve-Path $FolderEntry
+            $RP = Resolve-Path -LiteralPath $FolderEntry
             if ($RP.Provider.Name -ne 'FileSystem') {
                 Write-Error "$FunctionName provided path $FolderEntry is not on the FileSystem"
                 continue
@@ -263,7 +294,7 @@ function Global:Get-FolderAge {
             }
 
             if ($TestSubFolders) {
-                $FolderList = @(Get-ChildItem $FolderEntry -Directory -ea SilentlyContinue | Select -Expand FullName)
+                $FolderList = @(Get-ChildItem -LiteralPath $FolderEntry -Directory -ea SilentlyContinue | Select -Expand FullName)
                 if ($FolderList) {
                     Write-Verbose -Message "$(Get-Date -f T)   Processing $($FolderList.Count) subfolders of $FolderEntry"
                 } else {
@@ -280,16 +311,21 @@ function Global:Get-FolderAge {
                 
                 # processing single folder $Folder
 
+                if ($FoldersToSkip -and ($FoldersToSkip -contains $Folder)) {
+                    Write-Verbose -Message "$(Get-Date -f T)   skipping $Folder from processing, because it is present in $OutputFile."
+                    continue
+                }
+
                 if ($Threads -gt 1) {
 
                     $JobCode = "$FunctionName '$Folder'"
                     if ($CutOffTime) {$JobCode += " -CutOffTime '$CutOffString'"}
                     if ($Exclude) {$Join = "', '"; $JobCode += " -Exclude '$($Exclude -join $Join)'"}
                     if ($OutputFile) {$JobCode += " -OutputFile '$OutputFile'"}
-                    # TODO: Process other required parameters: CutOffTime, Exclude, OutputFile
                     Write-Verbose -Message "$(Get-Date -f T)   starting background job for '$Folder': $JobCode"
                     $JobCode = ". $SourceFile`n$JobCode" # first import function 
                     $JobList += Start-ThreadJob -ScriptBlock ([Scriptblock]::Create($JobCode)) -ThrottleLimit $Threads
+                    Start-Sleep -Milliseconds 200 # let the job execute begin block, before starting next one
 
                     continue # to next $Folder
                 }
@@ -298,7 +334,7 @@ function Global:Get-FolderAge {
                 $StartTime = Get-Date
                 $i = 0
                 $queue = @($Folder)
-                $LastWriteTime = Get-Item -Path $Folder | Select -Expand LastWriteTime
+                $LastWriteTime = Get-Item -LiteralPath $Folder | Select -Expand LastWriteTime
                 $TotalFiles = 0
                 $LastItemName = $Folder
                 $KeepProcessing = $true
@@ -391,29 +427,34 @@ function Global:Get-FolderAge {
                         LastWriteTime = $LastWriteTime
                         Modified = $Modified
                         Confident = $Confident
+                        # statistical info
                         TotalFiles = $TotalFiles
                         TotalFolders = $queue.Count
                         LastItem = $LastItemName
                         Depth = ($queue[$i-1].split($Separator)).Count - ($queue[0].split($Separator)).Count + 1
                         ElapsedSeconds = ($EndTime - $StartTime).TotalSeconds
                         FinishTime = $EndTime
+                        # error info
                         Errors = $ErrorsFound
                         LastError = $LastError
                     }
+                
+                #
                 # File output, if needed
+                #
+
                 if ($OutputFile) {
-                    if ($First) {
+                    if (!(Test-Path $OutputFile)) {
                         try {
                             $RetVal | Export-Csv -Path $OutputFile -Encoding Unicode -NoTypeInformation # Export-csv in PS v2 has no -LiteralPath
                             Write-Verbose -Message "$(Get-Date -f T)   created output file $OutputFile"
-                            $First = $false
                         } catch {
                             Write-Error "$FunctionName failed while writing to $OutputFile, file output is skipped`n$_"
                             $OutputFile = $null
                         }
                     } else {
                         try {
-                            $RetVal | ConvertTo-Csv -NoTypeInformation | Select -Skip 1 | Out-File -LiteralPath $OutputFile -Append -Encoding Unicode
+                            $RetVal | ConvertTo-Csv -NoTypeInformation | Select -Skip 1 | Out-File -FilePath $OutputFile -Append -Encoding Unicode
                             Write-Verbose -Message "$(Get-Date -f T)   appended new line to output file $OutputFile"
                         } catch {
                             Write-Error "$FunctionName failed to append date to $OutputFile, entry for $Folder will be skipped.`n$_"
@@ -431,7 +472,8 @@ function Global:Get-FolderAge {
         # if threads, receive them
         if ($Threads -gt 1) {
             Write-Verbose -Message "$(Get-Date -f T) $FunctionName waiting for background jobs results."
-            Receive-Job $JobList -Wait            
+            Receive-Job $JobList -Wait
+            Remove-Job $JobList
         }
 
         # function closing phase
@@ -439,3 +481,6 @@ function Global:Get-FolderAge {
     }
     
 }
+
+
+Set-Alias -Name gfa -Value Get-FolderAge